+++ test/PerpetualIntelligence.Terminal.Test/Extensions/IHostExtensionsRunConsoleRoutingTests.cs
@@ -9,7 +9,6 @@ using FluentAssertions;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
-using PerpetualIntelligence.Shared.Exceptions;
 using PerpetualIntelligence.Terminal.Commands.Handlers;
 using PerpetualIntelligence.Terminal.Commands.Routers;
 using PerpetualIntelligence.Terminal.Configuration.Options;
@@ -38,7 +37,7 @@ namespace PerpetualIntelligence.Terminal.Extensions
 
             // Set invalid start mode
             startContext = new TerminalStartContext(new TerminalStartInfo(TerminalStartMode.Grpc), tokenSource.Token);
-            Func<Task> act = async () => await host.RunConsoleRoutingAsync(new TerminalConsoleRoutingContext(startContext));
+            Func<Task> act = async () => await host.RunTerminalRoutingAsync<TerminalConsoleRouting, TerminalConsoleRoutingContext, TerminalConsoleRoutingResult>(new TerminalConsoleRoutingContext(startContext));
             await act.Should().ThrowAsync<TerminalException>().WithMessage("The requested start mode is not valid for console routing. start_mode=Grpc");
         }
 
@@ -57,7 +56,7 @@ namespace PerpetualIntelligence.Terminal.Extensions
             host = newHostBuilder.Build();
 
             GetCliOptions(host).Router.Timeout = Timeout.Infinite;
-            await host.RunConsoleRoutingAsync(new TerminalConsoleRoutingContext(startContext));
+            await host.RunTerminalRoutingAsync<TerminalConsoleRouting, TerminalConsoleRoutingContext, TerminalConsoleRoutingResult>(new TerminalConsoleRoutingContext(startContext));
 
             MockCommandRouter mockCommandRouter = (MockCommandRouter)host.Services.GetRequiredService<ICommandRouter>();
             mockCommandRouter.RouteCalled.Should().BeTrue();
@@ -85,7 +84,7 @@ namespace PerpetualIntelligence.Terminal.Extensions
             await Task.Delay(2050);
 
             GetCliOptions(host).Router.Timeout = Timeout.Infinite;
-            await host.RunConsoleRoutingAsync(new TerminalConsoleRoutingContext(startContext));
+            await host.RunTerminalRoutingAsync<TerminalConsoleRouting, TerminalConsoleRoutingContext, TerminalConsoleRoutingResult>(new TerminalConsoleRoutingContext(startContext));
 
             // Canceled task so router will not be called.
             MockCommandRouter mockCommandRouter = (MockCommandRouter)host.Services.GetRequiredService<ICommandRouter>();
@@ -115,7 +114,7 @@ namespace PerpetualIntelligence.Terminal.Extensions
             // Router will throw exception and then routing will get canceled
             GetCliOptions(host).Router.Timeout = Timeout.Infinite;
             GetCliOptions(host).Router.Caret = "$";
-            await host.RunConsoleRoutingAsync(new TerminalConsoleRoutingContext(startContext));
+            await host.RunTerminalRoutingAsync<TerminalConsoleRouting, TerminalConsoleRoutingContext, TerminalConsoleRoutingResult>(new TerminalConsoleRoutingContext(startContext));
 
             // Check the published error
             MockExceptionPublisher publisher = (MockExceptionPublisher)host.Services.GetRequiredService<IExceptionHandler>();
@@ -143,7 +142,7 @@ namespace PerpetualIntelligence.Terminal.Extensions
 
             // Router will throw exception and then routing will get canceled
             GetCliOptions(host).Router.Timeout = Timeout.Infinite;
-            await host.RunConsoleRoutingAsync(new TerminalConsoleRoutingContext(startContext));
+            await host.RunTerminalRoutingAsync<TerminalConsoleRouting, TerminalConsoleRoutingContext, TerminalConsoleRoutingResult>(new TerminalConsoleRoutingContext(startContext));
 
             // Check the published error
             MockExceptionPublisher publisher = (MockExceptionPublisher)host.Services.GetRequiredService<IExceptionHandler>();
@@ -175,7 +174,7 @@ namespace PerpetualIntelligence.Terminal.Extensions
 
             // Run the router for 5 seconds, the callback will stop the host 2 seconds.
             GetCliOptions(host).Router.Timeout = 5000;
-            await host.RunConsoleRoutingAsync(new TerminalConsoleRoutingContext(startContext));
+            await host.RunTerminalRoutingAsync<TerminalConsoleRouting, TerminalConsoleRoutingContext, TerminalConsoleRoutingResult>(new TerminalConsoleRoutingContext(startContext));
 
             // Till the timer callback cancel the route will be called multiple times.
             MockCommandRouter mockCommandRouter = (MockCommandRouter)host.Services.GetRequiredService<ICommandRouter>();
@@ -207,7 +206,7 @@ namespace PerpetualIntelligence.Terminal.Extensions
             // Router will throw exception and then routing will get canceled
             GetCliOptions(host).Router.Timeout = Timeout.Infinite;
             GetCliOptions(host).Router.Caret = ">$";
-            await host.RunConsoleRoutingAsync(new TerminalConsoleRoutingContext(startContext));
+            await host.RunTerminalRoutingAsync<TerminalConsoleRouting, TerminalConsoleRoutingContext, TerminalConsoleRoutingResult>(new TerminalConsoleRoutingContext(startContext));
 
             // Check the published error
             MockExceptionPublisher publisher = (MockExceptionPublisher)host.Services.GetRequiredService<IExceptionHandler>();
@@ -232,7 +231,7 @@ namespace PerpetualIntelligence.Terminal.Extensions
             // We will run in a infinite loop due to empty input so break that after 2 seconds
             tokenSource.CancelAfter(2000);
             GetCliOptions(host).Router.Timeout = Timeout.Infinite;
-            await host.RunConsoleRoutingAsync(new TerminalConsoleRoutingContext(startContext));
+            await host.RunTerminalRoutingAsync<TerminalConsoleRouting, TerminalConsoleRoutingContext, TerminalConsoleRoutingResult>(new TerminalConsoleRoutingContext(startContext));
 
             MockCommandRouter mockCommandRouter = (MockCommandRouter)host.Services.GetRequiredService<ICommandRouter>();
             mockCommandRouter.RouteCalled.Should().BeFalse();
@@ -256,7 +255,7 @@ namespace PerpetualIntelligence.Terminal.Extensions
             // send cancellation after 3 seconds. Idea is that in 3 seconds the router will route multiple times till canceled.
             tokenSource.CancelAfter(3000);
             GetCliOptions(host).Router.Timeout = Timeout.Infinite;
-            await host.RunConsoleRoutingAsync(new TerminalConsoleRoutingContext(startContext));
+            await host.RunTerminalRoutingAsync<TerminalConsoleRouting, TerminalConsoleRoutingContext, TerminalConsoleRoutingResult>(new TerminalConsoleRoutingContext(startContext));
 
             // In 3 seconds the Route will be called multiple times.
             MockCommandRouter mockCommandRouter = (MockCommandRouter)host.Services.GetRequiredService<ICommandRouter>();
@@ -285,7 +284,7 @@ namespace PerpetualIntelligence.Terminal.Extensions
             // Send cancellation after 2 seconds. Idea is in 2 seconds the router will route multiple times till canceled.
             tokenSource.CancelAfter(2000);
             GetCliOptions(host).Router.Timeout = Timeout.Infinite;
-            await host.RunConsoleRoutingAsync(new TerminalConsoleRoutingContext(startContext));
+            await host.RunTerminalRoutingAsync<TerminalConsoleRouting, TerminalConsoleRoutingContext, TerminalConsoleRoutingResult>(new TerminalConsoleRoutingContext(startContext));
 
             // Result is processed and disposed by handler not the routing service.
             mockCommandRouter.ReturnedRouterResult.Should().NotBeNull();
@@ -309,7 +308,7 @@ namespace PerpetualIntelligence.Terminal.Extensions
 
             // Route delay is set to 3000 and timeout is 2000
             GetCliOptions(host).Router.Timeout = 2000;
-            await host.RunConsoleRoutingAsync(new TerminalConsoleRoutingContext(startContext));
+            await host.RunTerminalRoutingAsync<TerminalConsoleRouting, TerminalConsoleRoutingContext, TerminalConsoleRoutingResult>(new TerminalConsoleRoutingContext(startContext));
 
             // Check the published error
             MockExceptionPublisher publisher = (MockExceptionPublisher)host.Services.GetRequiredService<IExceptionHandler>();
@@ -339,7 +338,7 @@ namespace PerpetualIntelligence.Terminal.Extensions
             tokenSource.CancelAfter(2000);
             GetCliOptions(host).Router.Timeout = Timeout.Infinite;
             GetCliOptions(host).Router.Caret = "test_caret";
-            await host.RunConsoleRoutingAsync(new TerminalConsoleRoutingContext(startContext));
+            await host.RunTerminalRoutingAsync<TerminalConsoleRouting, TerminalConsoleRoutingContext, TerminalConsoleRoutingResult>(new TerminalConsoleRoutingContext(startContext));
             titleWriter.ToString().Should().Be("test_caret");
 
             // Check output
@@ -366,7 +365,7 @@ namespace PerpetualIntelligence.Terminal.Extensions
             opt2.AddSingleton<IExceptionHandler>(new MockExceptionPublisher());
 
             // Add routing service
-            opt2.AddSingleton<TerminalConsoleRouting>();
+            opt2.AddSingleton<ITerminalRouting<TerminalConsoleRoutingContext, TerminalConsoleRoutingResult>, TerminalConsoleRouting>();
             opt2.AddSingleton<ITerminalConsole, TerminalSystemConsole>();
         }
 
@@ -385,7 +384,7 @@ namespace PerpetualIntelligence.Terminal.Extensions
             opt2.AddSingleton<IExceptionHandler>(new MockExceptionPublisher());
 
             // Add routing service
-            opt2.AddSingleton<TerminalConsoleRouting>();
+            opt2.AddSingleton<ITerminalRouting<TerminalConsoleRoutingContext, TerminalConsoleRoutingResult>, TerminalConsoleRouting>();
             opt2.AddSingleton<ITerminalConsole, TerminalSystemConsole>();
         }
 
@@ -404,7 +403,7 @@ namespace PerpetualIntelligence.Terminal.Extensions
             opt2.AddSingleton<IExceptionHandler>(new MockExceptionPublisher());
 
             // Add routing service
-            opt2.AddSingleton<TerminalConsoleRouting>();
+            opt2.AddSingleton<ITerminalRouting<TerminalConsoleRoutingContext, TerminalConsoleRoutingResult>, TerminalConsoleRouting>();
             opt2.AddSingleton<ITerminalConsole, TerminalSystemConsole>();
         }
 
@@ -423,7 +422,7 @@ namespace PerpetualIntelligence.Terminal.Extensions
             opt2.AddSingleton<IExceptionHandler>(new MockExceptionPublisher());
 
             // Add routing service
-            opt2.AddSingleton<TerminalConsoleRouting>();
+            opt2.AddSingleton<ITerminalRouting<TerminalConsoleRoutingContext, TerminalConsoleRoutingResult>, TerminalConsoleRouting>();
             opt2.AddSingleton<ITerminalConsole, TerminalSystemConsole>();
         }
 
@@ -443,7 +442,7 @@ namespace PerpetualIntelligence.Terminal.Extensions
             opt2.AddSingleton<IExceptionHandler>(new MockExceptionPublisher());
 
             // Add routing service
-            opt2.AddSingleton<TerminalConsoleRouting>();
+            opt2.AddSingleton<ITerminalRouting<TerminalConsoleRoutingContext, TerminalConsoleRoutingResult>, TerminalConsoleRouting>();
             opt2.AddSingleton<ITerminalConsole, TerminalSystemConsole>();
         }
 
@@ -463,7 +462,7 @@ namespace PerpetualIntelligence.Terminal.Extensions
             opt2.AddSingleton<IExceptionHandler>(new MockExceptionPublisher());
 
             // Add routing service
-            opt2.AddSingleton<TerminalConsoleRouting>();
+            opt2.AddSingleton<ITerminalRouting<TerminalConsoleRoutingContext, TerminalConsoleRoutingResult>, TerminalConsoleRouting>();
             opt2.AddSingleton<ITerminalConsole, TerminalSystemConsole>();
         }